fix: fix mapObejctDefs for tests
